change type of returned values from disk_space and update the function's description description

+++ src/linux/mod.rs
@@ -543,7 +543,7 @@ impl GeneralReadout for LinuxGeneralReadout {
         Err(ReadoutError::NotImplemented)
     }
 
-    fn disk_space(&self) -> Result<(u128, u128), ReadoutError> {
+    fn disk_space(&self) -> Result<(u64, u64), ReadoutError> {
         shared::disk_space(String::from("/"))
     }
